RandomGeneratorCommitFiles/share_producer/requirements.txt
kafka-python
pymongo
numpy
scipy
scikit-learn
python-dateutil
